refactor(archive): use writer semantic for batched updates
