+++ crates/redb/src/archive/mod.rs
@@ -7,9 +7,13 @@ use dolos_core::{
     ArchiveError, BlockBody, BlockSlot, ChainPoint, EraCbor, RawBlock, SlotTags, TxOrder,
 };
 
+use ::redb::Durability;
+use pallas::ledger::traverse::MultiEraBlock;
+use redb::WriteTransaction;
+use std::sync::Arc;
+
 mod indexes;
 mod tables;
-mod v1;
 
 #[derive(Debug)]
 pub struct RedbArchiveError(ArchiveError);
@@ -68,52 +72,9 @@ impl From<::redb::TransactionError> for RedbArchiveError {
 
 const DEFAULT_CACHE_SIZE_MB: usize = 500;
 
-fn compute_schema_hash(db: &Database) -> Result<Option<String>, RedbArchiveError> {
-    let mut hasher = pallas::crypto::hash::Hasher::<160>::new();
-
-    let rx = db.begin_read()?;
-
-    let names_1 = rx.list_tables()?.map(|t| t.name().to_owned());
-
-    let names_2 = rx.list_multimap_tables()?.map(|t| t.name().to_owned());
-
-    let mut names: Vec<_> = names_1.chain(names_2).collect();
-
-    debug!(tables = ?names, "tables names used to compute hash");
-
-    if names.is_empty() {
-        // this db hasn't been initialized, we can't compute hash
-        return Ok(None);
-    }
-
-    // sort to make sure we don't depend on some redb implementation regarding order
-    // of the tables.
-    names.sort();
-
-    names.into_iter().for_each(|n| hasher.input(n.as_bytes()));
-
-    let hash = hasher.finalize();
-
-    Ok(Some(hash.to_string()))
-}
-
-fn open_db(
-    path: impl AsRef<Path>,
-    cache_size: Option<usize>,
-) -> Result<Database, RedbArchiveError> {
-    let db = Database::builder()
-        .set_repair_callback(|x| warn!(progress = x.progress() * 100f64, "archive db is repairing"))
-        .set_cache_size(1024 * 1024 * cache_size.unwrap_or(DEFAULT_CACHE_SIZE_MB))
-        .create(path)?;
-
-    Ok(db)
-}
-
-const V1_HASH: &str = "b9e1e00f2427e24139929a60fbaf925948c11069";
-
 #[derive(Clone)]
-pub enum ChainStore {
-    SchemaV1(v1::ChainStore),
+pub struct ChainStore {
+    db: Arc<Database>,
 }
 
 impl ChainStore {
@@ -121,50 +82,54 @@ impl ChainStore {
         path: impl AsRef<Path>,
         cache_size: Option<usize>,
     ) -> Result<Self, RedbArchiveError> {
-        let db = open_db(path, cache_size)?;
-        let hash = compute_schema_hash(&db)?;
+        let db = Database::builder()
+            .set_repair_callback(|x| {
+                warn!(progress = x.progress() * 100f64, "archive db is repairing")
+            })
+            .set_cache_size(1024 * 1024 * cache_size.unwrap_or(DEFAULT_CACHE_SIZE_MB))
+            .create(path)?;
 
-        let schema = match hash.as_deref() {
-            // use stable schema if no hash
-            None => {
-                info!("no state db schema, initializing as v1");
-                v1::ChainStore::initialize(db)?.into()
-            }
-            Some(V1_HASH) => {
-                info!("detected state db schema v1");
-                v1::ChainStore::from(db).into()
-            }
-            Some(x) => panic!("can't recognize db hash {x}"),
-        };
-
-        Ok(schema)
+        Ok(Self::initialize(db)?)
     }
 
-    pub fn in_memory_v1() -> Result<Self, ArchiveError> {
+    pub fn in_memory() -> Result<Self, ArchiveError> {
         let db = ::redb::Database::builder()
             .create_with_backend(::redb::backends::InMemoryBackend::new())
             .map_err(RedbArchiveError::from)?;
 
-        let store = v1::ChainStore::initialize(db)?;
-        Ok(store.into())
+        Ok(Self::initialize(db)?)
+    }
+
+    pub fn initialize(db: Database) -> Result<Self, RedbArchiveError> {
+        let mut wx = db.begin_write()?;
+        wx.set_durability(Durability::Immediate);
+
+        indexes::Indexes::initialize(&wx)?;
+        tables::BlocksTable::initialize(&wx)?;
+
+        wx.commit()?;
+
+        Ok(Self { db: Arc::new(db) })
     }
 
     pub fn db(&self) -> &Database {
-        match self {
-            ChainStore::SchemaV1(x) => x.db(),
-        }
+        &self.db
     }
 
-    pub fn db_mut(&mut self) -> Option<&mut Database> {
-        match self {
-            ChainStore::SchemaV1(x) => x.db_mut(),
-        }
+    pub fn db_mut(&mut self) -> &mut Database {
+        Arc::get_mut(&mut self.db).unwrap()
     }
 
-    pub fn get_tip(&self) -> Result<Option<(BlockSlot, BlockBody)>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_tip(),
-        }
+    pub fn copy(&self, target: &Self) -> Result<(), RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        let wx = target.db().begin_write()?;
+
+        indexes::Indexes::copy(&rx, &wx)?;
+        tables::BlocksTable::copy(&rx, &wx)?;
+
+        wx.commit()?;
+
+        Ok(())
     }
 
     pub fn get_range(
@@ -172,120 +137,368 @@ impl ChainStore {
         from: Option<BlockSlot>,
         to: Option<BlockSlot>,
     ) -> Result<ChainRangeIter, RedbArchiveError> {
-        let out = match self {
-            ChainStore::SchemaV1(x) => x.get_range(from, to)?,
-        };
+        let rx = self.db().begin_read()?;
+        let range = tables::BlocksTable::get_range(&rx, from, to)?;
+        Ok(ChainRangeIter(range))
+    }
 
-        Ok(out)
+    pub fn start_writer(&self) -> Result<ChainStoreWriter, RedbArchiveError> {
+        let mut wx = self.db().begin_write()?;
+        wx.set_durability(Durability::Eventual);
+        wx.set_quick_repair(true);
+
+        Ok(ChainStoreWriter { wx })
     }
 
     pub fn find_intersect(
         &self,
         intersect: &[ChainPoint],
     ) -> Result<Option<ChainPoint>, RedbArchiveError> {
-        let out = match self {
-            ChainStore::SchemaV1(x) => x.find_intersect(intersect)?,
-        };
+        let rx = self.db().begin_read()?;
 
-        Ok(out)
+        for point in intersect {
+            let ChainPoint::Specific(slot, hash) = point else {
+                return Ok(Some(ChainPoint::Origin));
+            };
+
+            if let Some(body) = tables::BlocksTable::get_by_slot(&rx, *slot)? {
+                let decoded =
+                    MultiEraBlock::decode(&body).map_err(ArchiveError::BlockDecodingError)?;
+
+                if decoded.hash().eq(hash) {
+                    return Ok(Some(ChainPoint::Specific(decoded.slot(), decoded.hash())));
+                }
+            }
+        }
+
+        Ok(None)
     }
 
-    pub fn get_block_by_hash(
+    pub fn get_possible_block_slots_by_address_payment_part(
+        &self,
+        address_payment_part: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_address_payment_part(&rx, address_payment_part)
+    }
+
+    pub fn get_possible_block_slots_by_address_stake_part(
+        &self,
+        address_stake_part: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_address_stake_part(&rx, address_stake_part)
+    }
+
+    pub fn get_possible_block_slots_by_asset(
+        &self,
+        asset: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_asset(&rx, asset)
+    }
+
+    pub fn get_possible_block_slots_by_block_hash(
         &self,
         block_hash: &[u8],
-    ) -> Result<Option<BlockBody>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_block_by_hash(block_hash),
-        }
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_block_hash(&rx, block_hash)
     }
 
-    pub fn get_block_by_slot(
+    pub fn get_possible_block_slots_by_block_number(
         &self,
-        slot: &BlockSlot,
-    ) -> Result<Option<BlockBody>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_block_by_slot(slot),
-        }
+        block_number: &u64,
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_block_number(&rx, block_number)
     }
 
-    pub fn get_block_by_number(&self, number: &u64) -> Result<Option<BlockBody>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_block_by_number(number),
-        }
+    pub fn get_possible_block_slots_by_datum_hash(
+        &self,
+        datum_hash: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_datum_hash(&rx, datum_hash)
+    }
+
+    pub fn get_possible_block_slots_by_policy(
+        &self,
+        policy: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_policy(&rx, policy)
+    }
+
+    pub fn get_possible_block_slots_by_script_hash(
+        &self,
+        script_hash: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_script_hash(&rx, script_hash)
+    }
+
+    pub fn get_possible_block_slots_by_tx_hash(
+        &self,
+        tx_hash: &[u8],
+    ) -> Result<Vec<BlockSlot>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        indexes::Indexes::get_by_tx_hash(&rx, tx_hash)
+    }
+
+    pub fn get_possible_blocks_by_address_payment_part(
+        &self,
+        address_payment_part: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_address_payment_part(address_payment_part)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_address_stake_part(
+        &self,
+        address_stake_part: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_address_stake_part(address_stake_part)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_asset(
+        &self,
+        asset: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_asset(asset)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_block_hash(
+        &self,
+        block_hash: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_block_hash(block_hash)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_block_number(
+        &self,
+        block_number: &u64,
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_block_number(block_number)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_datum_hash(
+        &self,
+        datum_hash: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_datum_hash(datum_hash)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_policy(
+        &self,
+        policy: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_policy(policy)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_script_hash(
+        &self,
+        script_hash: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_script_hash(script_hash)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
+    }
+
+    pub fn get_possible_blocks_by_tx_hash(
+        &self,
+        tx_hash: &[u8],
+    ) -> Result<Vec<BlockBody>, RedbArchiveError> {
+        self.get_possible_block_slots_by_tx_hash(tx_hash)?
+            .iter()
+            .flat_map(|slot| match self.get_block_by_slot(slot) {
+                Ok(Some(block)) => Some(Ok(block)),
+                Ok(None) => None,
+                Err(e) => Some(Err(e)),
+            })
+            .collect()
     }
 
     pub fn get_block_with_tx(
         &self,
         tx_hash: &[u8],
     ) -> Result<Option<(BlockBody, TxOrder)>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_block_with_tx(tx_hash),
+        let possible = self.get_possible_blocks_by_tx_hash(tx_hash)?;
+
+        for raw in possible {
+            let block =
+                MultiEraBlock::decode(raw.as_slice()).map_err(ArchiveError::BlockDecodingError)?;
+            for (idx, tx) in block.txs().iter().enumerate() {
+                if tx.hash().to_vec() == tx_hash {
+                    return Ok(Some((raw, idx)));
+                }
+            }
         }
+
+        Ok(None)
     }
 
-    pub fn iter_blocks_with_address(
+    pub fn iter_possible_blocks_with_address(
         &self,
         address: &[u8],
     ) -> Result<ChainSparseIter, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => {
-                // TODO: we need to filter the false positives
-                x.iter_possible_blocks_with_address(address)
-            }
-        }
+        let rx = self.db().begin_read()?;
+        let range = indexes::Indexes::iter_by_address(&rx, address)?;
+        Ok(ChainSparseIter(rx, range))
     }
 
-    pub fn iter_blocks_with_asset(
+    pub fn iter_possible_blocks_with_asset(
         &self,
         asset: &[u8],
     ) -> Result<ChainSparseIter, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => {
-                // TODO: we need to filter the false positives
-                x.iter_possible_blocks_with_asset(asset)
-            }
-        }
+        let rx = self.db().begin_read()?;
+        let range = indexes::Indexes::iter_by_asset(&rx, asset)?;
+        Ok(ChainSparseIter(rx, range))
     }
 
-    pub fn iter_blocks_with_payment(
+    pub fn iter_possible_blocks_with_payment(
         &self,
         payment: &[u8],
     ) -> Result<ChainSparseIter, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => {
-                // TODO: we need to filter the false positives
-                x.iter_possible_blocks_with_payment(payment)
+        let rx = self.db().begin_read()?;
+        let range = indexes::Indexes::iter_by_payment(&rx, payment)?;
+        Ok(ChainSparseIter(rx, range))
+    }
+
+    pub fn get_block_by_slot(
+        &self,
+        slot: &BlockSlot,
+    ) -> Result<Option<BlockBody>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        tables::BlocksTable::get_by_slot(&rx, *slot)
+    }
+
+    pub fn get_block_by_hash(
+        &self,
+        block_hash: &[u8],
+    ) -> Result<Option<BlockBody>, RedbArchiveError> {
+        let possible: Vec<BlockBody> = self.get_possible_blocks_by_block_hash(block_hash)?;
+
+        for raw in possible {
+            let block =
+                MultiEraBlock::decode(raw.as_slice()).map_err(ArchiveError::BlockDecodingError)?;
+
+            if block.hash().as_slice() == block_hash {
+                return Ok(Some(raw));
             }
+        }
+
+        Ok(None)
+    }
+
+    pub fn get_block_by_number(
+        &self,
+        block_number: &u64,
+    ) -> Result<Option<BlockBody>, RedbArchiveError> {
+        let possible = self.get_possible_blocks_by_block_number(block_number)?;
+
+        for raw in possible {
+            let block =
+                MultiEraBlock::decode(raw.as_slice()).map_err(ArchiveError::BlockDecodingError)?;
+
+            if block.number() == *block_number {
+                return Ok(Some(raw));
+            }
+        }
+
+        Ok(None)
+    }
+
+    pub fn get_slot_for_tx(&self, tx_hash: &[u8]) -> Result<Option<BlockSlot>, RedbArchiveError> {
+        let mut possible = self.get_possible_block_slots_by_tx_hash(tx_hash)?;
+        if possible.len() == 1 {
+            Ok(possible.pop())
+        } else {
+            for slot in possible {
+                let block = self.get_block_by_slot(&slot)?;
+
+                if let Some(raw) = block {
+                    let block =
+                        MultiEraBlock::decode(&raw).map_err(ArchiveError::BlockDecodingError)?;
+                    if block.txs().iter().any(|x| x.hash().to_vec() == tx_hash) {
+                        return Ok(Some(slot));
+                    }
+                }
+            }
+            Ok(None)
         }
     }
 
     pub fn get_tx(&self, tx_hash: &[u8]) -> Result<Option<EraCbor>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_tx(tx_hash),
+        let possible = self.get_possible_blocks_by_tx_hash(tx_hash)?;
+
+        for raw in possible {
+            let block =
+                MultiEraBlock::decode(raw.as_slice()).map_err(ArchiveError::BlockDecodingError)?;
+            if let Some(tx) = block.txs().iter().find(|x| x.hash().to_vec() == tx_hash) {
+                return Ok(Some(EraCbor(block.era().into(), tx.encode())));
+            }
         }
+
+        Ok(None)
     }
 
-    pub fn get_slot_for_tx(&self, tx_hash: &[u8]) -> Result<Option<BlockSlot>, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.get_slot_for_tx(tx_hash),
-        }
-    }
-
-    pub fn apply(
-        &self,
-        point: &ChainPoint,
-        block: &RawBlock,
-        tags: &SlotTags,
-    ) -> Result<(), RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => Ok(x.apply(point, block, tags)?),
-        }
-    }
-
-    pub fn undo(&self, point: &ChainPoint, tags: &SlotTags) -> Result<(), RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => Ok(x.undo(point, tags)?),
-        }
+    pub fn get_tip(&self) -> Result<Option<(BlockSlot, BlockBody)>, RedbArchiveError> {
+        let rx = self.db().begin_read()?;
+        tables::BlocksTable::get_tip(&rx)
     }
 
     pub fn prune_history(
@@ -293,15 +506,103 @@ impl ChainStore {
         max_slots: u64,
         max_prune: Option<u64>,
     ) -> Result<bool, RedbArchiveError> {
-        match self {
-            ChainStore::SchemaV1(x) => x.prune_history(max_slots, max_prune),
+        let rx = self.db().begin_read()?;
+        let start = match tables::BlocksTable::first(&rx)? {
+            Some((slot, _)) => slot,
+            None => {
+                debug!("no start point found on chain, skipping housekeeping");
+                return Ok(true);
+            }
+        };
+
+        let last = match tables::BlocksTable::last(&rx)? {
+            Some((slot, _)) => slot,
+            None => {
+                debug!("no tip found on chain, skipping housekeeping");
+                return Ok(true);
+            }
+        };
+
+        let delta = last.saturating_sub(start);
+        let excess = delta.saturating_sub(max_slots);
+
+        debug!(delta, excess, last, start, "chain history delta computed");
+
+        if excess == 0 {
+            debug!(delta, max_slots, excess, "no pruning necessary on chain");
+            return Ok(true);
         }
+
+        let (done, max_prune) = match max_prune {
+            Some(max) => (excess <= max, core::cmp::min(excess, max)),
+            None => (true, excess),
+        };
+
+        let prune_before = start + max_prune;
+
+        info!(
+            cutoff_slot = prune_before,
+            start, excess, "pruning archive for excess history"
+        );
+
+        let mut wx = self.db().begin_write()?;
+        wx.set_quick_repair(true);
+
+        tables::BlocksTable::remove_before(&wx, prune_before)?;
+
+        wx.commit()?;
+
+        Ok(done)
+    }
+}
+
+impl From<Database> for ChainStore {
+    fn from(value: Database) -> Self {
+        Self {
+            db: Arc::new(value),
+        }
+    }
+}
+
+pub struct ChainStoreWriter {
+    wx: WriteTransaction,
+}
+
+impl dolos_core::ArchiveWriter for ChainStoreWriter {
+    fn apply(
+        &self,
+        point: &ChainPoint,
+        block: &RawBlock,
+        tags: &SlotTags,
+    ) -> Result<(), ArchiveError> {
+        indexes::Indexes::apply(&self.wx, point, tags)?;
+        tables::BlocksTable::apply(&self.wx, point, block)?;
+
+        Ok(())
+    }
+
+    fn undo(&self, point: &ChainPoint, tags: &SlotTags) -> Result<(), ArchiveError> {
+        indexes::Indexes::undo(&self.wx, point, tags)?;
+        tables::BlocksTable::undo(&self.wx, point)?;
+
+        Ok(())
+    }
+
+    fn commit(self) -> Result<(), ArchiveError> {
+        self.wx.commit().map_err(RedbArchiveError::from)?;
+
+        Ok(())
     }
 }
 
 impl dolos_core::ArchiveStore for ChainStore {
     type BlockIter<'a> = ChainRangeIter;
     type SparseBlockIter = ChainSparseIter;
+    type Writer = ChainStoreWriter;
+
+    fn start_writer(&self) -> Result<Self::Writer, ArchiveError> {
+        Ok(Self::start_writer(self)?)
+    }
 
     fn get_block_by_hash(&self, block_hash: &[u8]) -> Result<Option<BlockBody>, ArchiveError> {
         Ok(Self::get_block_by_hash(self, block_hash)?)
@@ -334,18 +635,27 @@ impl dolos_core::ArchiveStore for ChainStore {
         &self,
         address: &[u8],
     ) -> Result<Self::SparseBlockIter, ArchiveError> {
-        Ok(Self::iter_blocks_with_address(self, address)?)
+        // TODO: we need to filter the false positives
+        let out = self.iter_possible_blocks_with_address(address)?;
+
+        Ok(out)
     }
 
     fn iter_blocks_with_asset(&self, asset: &[u8]) -> Result<Self::SparseBlockIter, ArchiveError> {
-        Ok(Self::iter_blocks_with_asset(self, asset)?)
+        // TODO: we need to filter the false positives
+        let out = self.iter_possible_blocks_with_asset(asset)?;
+
+        Ok(out)
     }
 
     fn iter_blocks_with_payment(
         &self,
         payment: &[u8],
     ) -> Result<Self::SparseBlockIter, ArchiveError> {
-        Ok(Self::iter_blocks_with_payment(self, payment)?)
+        // TODO: we need to filter the false positives
+        let out = self.iter_possible_blocks_with_payment(payment)?;
+
+        Ok(out)
     }
 
     fn get_range<'a>(
@@ -364,27 +674,8 @@ impl dolos_core::ArchiveStore for ChainStore {
         Ok(Self::get_tip(self)?)
     }
 
-    fn apply(
-        &self,
-        point: &ChainPoint,
-        block: &RawBlock,
-        tags: &SlotTags,
-    ) -> Result<(), ArchiveError> {
-        Ok(Self::apply(self, point, block, tags)?)
-    }
-
-    fn undo(&self, point: &ChainPoint, tags: &SlotTags) -> Result<(), ArchiveError> {
-        Ok(Self::undo(self, point, tags)?)
-    }
-
     fn prune_history(&self, max_slots: u64, max_prune: Option<u64>) -> Result<bool, ArchiveError> {
         Ok(Self::prune_history(self, max_slots, max_prune)?)
-    }
-}
-
-impl From<v1::ChainStore> for ChainStore {
-    fn from(value: v1::ChainStore) -> Self {
-        Self::SchemaV1(value)
     }
 }
 
@@ -447,17 +738,5 @@ impl DoubleEndedIterator for ChainSparseIter {
         };
 
         Some(Ok((slot, block)))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn schema_hash_computation() {
-        let store = ChainStore::in_memory_v1().unwrap();
-        let hash = compute_schema_hash(store.db()).unwrap();
-        assert_eq!(hash.unwrap(), V1_HASH);
     }
 }
